[Env][docker.intel_oneapi] Fix NaNs emitted by random gen in test (#910)

+++ env/machine/docker/intel_oneapi/env_built_intel-llvm.sh
@@ -1,8 +1,13 @@
 # Exports will be provided by the new env script above this line
 # will be exported : ACPP_GIT_DIR, ACPP_BUILD_DIR, ACPP_INSTALL_DIR
 
-apt update
-apt install -y python3-dev
+# test if python3-dev installed
+# dpkg -l python3-dev
+if ! dpkg -l python3.*-dev &> /dev/null; then
+    echo "python3-dev is not installed. Installing it."
+    apt update
+    apt install -y python3-dev
+fi
 
 function shamconfigure {
     cmake \
@@ -12,7 +17,7 @@ function shamconfigure {
         -DSYCL_IMPLEMENTATION=IntelLLVM \
         -DINTEL_LLVM_PATH=$(dirname $(which icpx))/.. \
         -DCMAKE_CXX_COMPILER=$(which icpx) \
-        -DCMAKE_CXX_FLAGS="-fsycl" \
+        -DCMAKE_CXX_FLAGS="-fsycl -fp-model=precise" \
         -DCMAKE_BUILD_TYPE="${SHAMROCK_BUILD_TYPE}" \
         -DBUILD_TEST=Yes \
         "${CMAKE_OPT[@]}"
